fix: disabled rag while in-development
